Integrated webui with rest-services in preparation for automated heroku deployment
