--- demo/src/main/java/com/example/demo/ProductRepo.java
@@ -1,6 +1,0 @@
-package com.example.demo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepo extends JpaRepository<Product,Integer> {
-}
