--- demo/src/main/java/com/example/demo/StateRepo.java
@@ -1,10 +1,0 @@
-package com.example.demo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface StateRepo extends JpaRepository<State,Integer> {
-    List<State> findByName(String state);
-
-}
