--- demo/src/main/resources/application.properties
@@ -1,9 +1,0 @@
-
-spring.datasource.driverClassName=org.postgresql.Driver
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto = update
-spring.datasource.url=jdbc:postgresql://ec2-3-222-150-253.compute-1.amazonaws.com/d9qbgg7j8r8m6
-spring.datasource.username=rtivwuhyxbakzy
-spring.datasource.password=59562cf29aa5d2806c4ecabb4b8c9cdd8f1b72fb643d2a87adebbcf99dd6c788
-spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
-
