+++ frontend/package.json
@@ -19,11 +19,6 @@
     "node": "12.16.x",
     "npm": "6.x"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "npm run lint"
-    }
-  },
   "private": true,
   "dependencies": {
     "@angular/animations": "~9.1.1",
