+++ src/main/java/org/emeritosbanditos/backend/ProductRepo.java
@@ -1,0 +1,6 @@
+package org.emeritosbanditos.backend;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepo extends JpaRepository<Product,Integer> {
+}
