+++ src/main/java/org/emeritosbanditos/backend/StateRepo.java
@@ -1,0 +1,10 @@
+package org.emeritosbanditos.backend;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface StateRepo extends JpaRepository<State,Integer> {
+    List<State> findByName(String state);
+
+}
